Remove work around for lingering DHCID records
Kea ddns used the same TSIG key for forward and
reverse DNS zones. Even though the content and
name of the keys was identical, it appears to have
been the cause of a DHCID record not being released
when the cooresponding A record was released.
Forward and reverse zones use the same TSIG key.